Remove redundand classes
Класс IGitHubWrapper на самом деле был фасадом, а IGitHubFacade вообще
не нужен, его функции выполняет GiraProjectsDirectory.
