+++ Source/Model/GiraProjects/GitHubSettings.cs
