+++ Source/Model/GitHubFacades/Installation.cs
