+++ Source/Model/GitHubFacades/RepositoryProjectModel.cs
@@ -1,6 +1,6 @@
 ﻿namespace Model
 {
-	internal class ProjectModel
+	internal class RepositoryProjectModel
 	{
 		public int Id { get; set; }
