--- Source/Model/Repos/ReposDataSource.cs
@@ -1,44 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace Model
-{
-	public class ReposDataSource
-	{
-		private IGitHubWrapper GitHub { get; }
-
-		public ReposDataSource(IGitHubWrapper gitHub)
-		{
-			GitHub = gitHub;
-		}
-
-		public RepoSettings GetRepoSettings(long installationId, long repoId)
-		{
-			var repo = GitHub.GetRepository(installationId, repoId);
-
-			// todo: Здесь должны считываться настройки репозитория из БД.
-			return new RepoSettings(
-				repo: repo,
-				new DateTime(2021, 02, 8),
-				sprintDuration: 7,
-				new[] {
-					new Lane(0, "Bugs", "bug"),
-					new Lane(1, "Tech Debts", "tech debt"),
-					new Lane(2, "Road Map", "road map")
-				},
-				new[] { new RepositoryProject(1, "Caprica2.0") });
-		}
-
-		public IReadOnlyCollection<Issue> GetRepoIssues(long installationId, long repoId)
-		{
-			var repoSettings = GetRepoSettings(installationId, repoId);
-
-			var issues = GitHub.GetRepositoryIssues(installationId, repoId);
-
-			return issues
-				.Where(issue => issue.Projects.ContainsAny(repoSettings.AllowedProjects))
-				.ToArray();
-		}
-	}
-}
