--- Source/WebAPI/Installations/InstallationDto.cs
@@ -1,19 +1,0 @@
-﻿using Model;
-
-namespace WebAPI
-{
-	public class InstallationDto
-	{
-		public long Id { get; set; }
-
-		public string OwnerLogin { get; set; }
-
-		public static InstallationDto BuildFrom(Installation installation)
-		{
-			return new InstallationDto {
-				Id = installation.Id,
-				OwnerLogin = installation.OwnerLogin
-			};
-		}
-	}
-}
