+++ Source/WebAPI/Projects/GiraProjectsController.cs
@@ -7,20 +7,21 @@ namespace WebAPI
 {
 	[Authorize]
 	[ApiController]
-	[Route("[controller]")]
+	[Route("projects")]
 	public class GiraProjectsController : Controller
 	{
-		private IGitHubFacade Facade { get; }
+		private GiraProjectsDirectory Directory { get; }
 
-		public GiraProjectsController(IGitHubFacade facade)
+		public GiraProjectsController(GiraProjectsDirectory directory)
 		{
-			Facade = facade;
+			Directory = directory;
 		}
 
 		[HttpGet]
+		[Route("{projectId}")]
 		public GiraProjectDto GetGiraProject(int projectId)
 		{
-			var giraProject = Facade.GetGiraProject(projectId);
+			var giraProject = Directory.GetGiraProject(projectId);
 
 			return new GiraProjectDto {
 				Id = giraProject.Id,
@@ -46,18 +47,31 @@ namespace WebAPI
 		}
 
 		[HttpGet]
-		[Route("tasks")]
+		[Route("{projectId}/tasks")]
 		public GiraTaskDto[] GetGiraProjectTasks(int projectId)
 		{
-			var giraProject = Facade.GetGiraProject(projectId);
-
-			return giraProject.Tasks
+			return Directory
+				.GetGiraProjectTasks(projectId)
 				.Select(giraTask => new GiraTaskDto {
 					Id = giraTask.Id,
 					ClosedAt = giraTask.ClosedAt,
 					CreatedAt = giraTask.CreatedAt,
 					Title = giraTask.Title,
 					Labels = giraTask.Labels.ToArray()
+				})
+				.ToArray();
+		}
+
+		[HttpGet]
+		[Route("{projectId}/sprints")]
+		public SprintDto[] GetProjectSprints(int projectId)
+		{
+			return Directory
+				.GetGiraProjectSprints(projectId)
+				.Select(sprint => new SprintDto {
+					Id = sprint.Id,
+					BeginAt = sprint.BeginAt,
+					Duration = sprint.Duration
 				})
 				.ToArray();
 		}
